only attempt env.js copy in docker-entrypoint.sh if file exists

+++ docker-entrypoint.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
+ENV_FILE_PATH=/env.js
 
-exec cp /env.js /usr/share/nginx/www/assets/env.js
+if [ -e ENV_FILE_PATH ]; then
+  exec cp /env.js /usr/share/nginx/www/assets/env.js
+fi
+
 exec nginx -g 'daemon off;' "$@"
